adjust sync result returned directly + support response=raw for single outputs or multi-by-ref (relates to #376)
