+++ tests/functional/test_builtin.py
@@ -86,30 +86,80 @@ class BuiltinAppTest(WpsConfigBase):
         assert body["jobControlOptions"] == [ExecuteControlOption.ASYNC, ExecuteControlOption.SYNC]
         assert body["outputTransmission"] == [ExecuteTransmissionMode.REFERENCE, ExecuteTransmissionMode.VALUE]
 
-    def test_jsonarray2netcdf_execute_async(self):
+    def setup_inputs(self, stack):
         dirname = tempfile.gettempdir()
         nc_data = "Hello NetCDF!"
+        tmp_ncdf = tempfile.NamedTemporaryFile(dir=dirname, mode="w", suffix=".nc")
+        tmp_json = tempfile.NamedTemporaryFile(dir=dirname, mode="w", suffix=".json")
+        tmp_ncdf = stack.enter_context(tmp_ncdf)  # noqa
+        tmp_json = stack.enter_context(tmp_json)  # noqa
+        tmp_ncdf.write(nc_data)
+        tmp_ncdf.seek(0)
+        tmp_json.write(json.dumps(["file://{}".format(os.path.join(dirname, tmp_ncdf.name))]))
+        tmp_json.seek(0)
+        body = {"inputs": [{"id": "input", "href": os.path.join(dirname, tmp_json.name)}]}
+        return body, nc_data
+
+    def validate_results(self, results, outputs, data, links):
+        # first validate format of OGC-API results
+        if results is not None:
+            assert isinstance(results, dict)
+            assert "output" in results, "Expected result ID 'output' in response body"
+            assert isinstance(results["output"], dict), "Container of result ID 'output' should be a dict"
+            assert "href" in results["output"]
+            assert "format" in results["output"]
+            fmt = results["output"]["format"]  # type: JSON
+            assert isinstance(fmt, dict), "Result format should be provided with content details"
+            assert "mediaType" in fmt
+            assert isinstance(fmt["mediaType"], str), "Result format Content-Type should be a single string definition"
+            assert fmt["mediaType"] == ContentType.APP_NETCDF, "Result 'output' format expected to be NetCDF file"
+            nc_href = results["output"]["href"]
+            assert isinstance(nc_href, str) and len(nc_href)
+        elif links:
+            assert isinstance(links, list) and len(links) == 1 and isinstance(links[0], tuple)
+            assert "rel=\"output\"" in links[0][1]
+            assert f"type={ContentType.APP_NETCDF}" in links[0][1]
+            nc_link = links[0][1].split(" ")[0]
+            assert nc_link.startswith("<") and nc_link.startswith(">")
+            nc_href = nc_link[1:-1]
+        else:
+            nc_href = None
+
+        settings = get_settings_from_testapp(self.app)
+        wps_path = settings.get("weaver.wps_output_path")
+        wps_dir = settings.get("weaver.wps_output_dir")
+        wps_out = "{}{}".format(settings.get("weaver.url"), wps_path)
+
+        # validate results if applicable
+        if nc_href is not None:
+            nc_real_path = nc_href.replace(wps_out, wps_dir)
+            assert nc_href.startswith(wps_out)
+            assert os.path.split(nc_real_path)[-1] == os.path.split(nc_href)[-1]
+            assert os.path.isfile(nc_real_path)
+            with open(nc_real_path, "r") as f:
+                assert f.read() == data
+
+        # if everything was valid for results, validate equivalent but differently formatted outputs response
+        assert outputs["outputs"][0]["id"] == "output"
+        nc_href = outputs["outputs"][0]["href"]
+        assert isinstance(nc_href, str) and len(nc_href)
+        assert nc_href.startswith(wps_out)
+        nc_real_path = nc_href.replace(wps_out, wps_dir)
+        assert os.path.split(nc_real_path)[-1] == os.path.split(nc_href)[-1]
+
+    def test_jsonarray2netcdf_execute_async(self):
         with contextlib.ExitStack() as stack_exec:
-            tmp_ncdf = tempfile.NamedTemporaryFile(dir=dirname, mode="w", suffix=".nc")
-            tmp_json = tempfile.NamedTemporaryFile(dir=dirname, mode="w", suffix=".json")
-            tmp_ncdf = stack_exec.enter_context(tmp_ncdf)  # noqa
-            tmp_json = stack_exec.enter_context(tmp_json)  # noqa
-            tmp_ncdf.write(nc_data)
-            tmp_ncdf.seek(0)
-            tmp_json.write(json.dumps(["file://{}".format(os.path.join(dirname, tmp_ncdf.name))]))
-            tmp_json.seek(0)
-            data = {
+            body, nc_data = self.setup_inputs(stack_exec)
+            body.update({
                 "mode": ExecuteMode.ASYNC,
                 "response": ExecuteResponse.DOCUMENT,
-                "inputs": [{"id": "input", "href": os.path.join(dirname, tmp_json.name)}],
                 "outputs": [{"id": "output", "transmissionMode": ExecuteTransmissionMode.VALUE}],
-            }
-
+            })
             for mock_exec in mocked_execute_celery():
                 stack_exec.enter_context(mock_exec)
             path = "/processes/jsonarray2netcdf/jobs"
             resp = mocked_sub_requests(self.app, "post_json", path,
-                                       data=data, headers=self.json_headers, only_local=True)
+                                       data=body, headers=self.json_headers, only_local=True)
 
         assert resp.status_code == 201, "Error: {}".format(resp.json)
         assert resp.content_type in ContentType.APP_JSON
@@ -127,48 +177,180 @@ class BuiltinAppTest(WpsConfigBase):
         assert resp.status_code == 200, "Error job outputs:\n{}".format(resp.json)
         outputs = resp.json
 
-        self.validate_results(results, outputs, nc_data)
+        self.validate_results(results, outputs, nc_data, None)
 
-    def test_jsonarray2netcdf_execute_sync(self):
-        dirname = tempfile.gettempdir()
-        nc_data = "Hello NetCDF!"
+    def test_jsonarray2netcdf_execute_async_output_by_reference_dontcare_response_document(self):
+        """
+        Jobs submitted with ``response=document`` are not impacted by ``transmissionMode``.
+
+        The results schema should always be returned when document is requested.
+
+        .. seealso::
+            https://docs.ogc.org/is/18-062r2/18-062r2.html#req_core_process-execute-sync-document
+        """
         with contextlib.ExitStack() as stack_exec:
-            tmp_ncdf = tempfile.NamedTemporaryFile(dir=dirname, mode="w", suffix=".nc")
-            tmp_json = tempfile.NamedTemporaryFile(dir=dirname, mode="w", suffix=".json")
-            tmp_ncdf = stack_exec.enter_context(tmp_ncdf)  # noqa
-            tmp_json = stack_exec.enter_context(tmp_json)  # noqa
-            tmp_ncdf.write(nc_data)
-            tmp_ncdf.seek(0)
-            tmp_json.write(json.dumps(["file://{}".format(os.path.join(dirname, tmp_ncdf.name))]))
-            tmp_json.seek(0)
-            data = {
-                "inputs": [{"id": "input", "href": os.path.join(dirname, tmp_json.name)}],
-                "outputs": [{"id": "output", "transmissionMode": ExecuteTransmissionMode.VALUE}],
-            }
-            headers = {"Prefer": "wait=10"}
-            headers.update(self.json_headers)
-
+            body, nc_data = self.setup_inputs(stack_exec)
+            body.update({
+                "response": ExecuteResponse.DOCUMENT,  # by value/reference don't care because of this
+                "outputs": [{"id": "output", "transmissionMode": ExecuteTransmissionMode.REFERENCE}],
+            })
             for mock_exec in mocked_execute_celery():
                 stack_exec.enter_context(mock_exec)
             path = "/processes/jsonarray2netcdf/jobs"
             resp = mocked_sub_requests(self.app, "post_json", path,
-                                       data=data, headers=headers, only_local=True)
+                                       data=body, headers=self.json_headers, only_local=True)
+
+        assert resp.status_code == 201, "Error: {}".format(resp.json)
+        assert resp.content_type in ContentType.APP_JSON
+        job_url = resp.json["location"]
+        self.monitor_job(job_url, return_status=True)  # don't fetch results automatically
+
+        resp = self.app.get("{}/results".format(job_url), headers=self.json_headers)
+        assert resp.status_code == 200, "Error: {}".format(resp.text)
+        assert resp.content_type == ContentType.APP_JSON
+        result_links = [hdr for hdr in resp.headers if hdr[0].lower() == "link"]
+        assert len(result_links) == 0
+        results = resp.json
+
+        # even though results are requested by Link reference,
+        # Weaver still offers them with document on outputs endpoint
+        output_url = job_url + "/outputs"
+        resp = self.app.get(output_url, headers=self.json_headers)
+        assert resp.status_code == 200, "Error job outputs:\n{}".format(resp.text)
+        outputs = resp.json
+
+        self.validate_results(results, outputs, nc_data, result_links)
+
+    def test_jsonarray2netcdf_execute_async_output_by_value_response_raw(self):
+        """
+        Jobs submitted with ``response=raw`` and single output as ``transmissionMode=value`` must return its raw data.
+
+        .. seealso::
+            https://docs.ogc.org/is/18-062r2/18-062r2.html#req_core_process-execute-sync-raw-value-one
+        """
+        with contextlib.ExitStack() as stack_exec:
+            body, nc_data = self.setup_inputs(stack_exec)
+            body.update({
+                "response": ExecuteResponse.RAW,  # by value/reference important here
+                # NOTE: quantity of outputs important as well
+                #       since single output, content-type is directly that output (otherwise should be multipart)
+                "outputs": [{"id": "output", "transmissionMode": ExecuteTransmissionMode.VALUE}],  # data dump
+            })
+            for mock_exec in mocked_execute_celery():
+                stack_exec.enter_context(mock_exec)
+            path = "/processes/jsonarray2netcdf/jobs"
+            resp = mocked_sub_requests(self.app, "post_json", path,
+                                       data=body, headers=self.json_headers, only_local=True)
+
+        assert resp.status_code == 201, "Error: {}".format(resp.text)
+        assert resp.content_type in ContentType.APP_JSON
+        job_url = resp.json["location"]
+        self.monitor_job(job_url, return_status=True)  # don't fetch results automatically
+
+        resp = self.app.get("{}/results".format(job_url), headers=self.json_headers)
+        assert resp.status_code < 400, "Error: {}".format(resp.text)
+        assert resp.status_code == 200, "Body should contain literal raw data dump"
+        assert resp.content_type in ContentType.APP_NETCDF, "raw result by value should be directly the content-type"
+        assert resp.text == nc_data, "raw result by value should be directly the data content"
+        assert resp.headers
+        result_links = [hdr for hdr in resp.headers if hdr[0].lower() == "link"]
+        assert len(result_links) == 0
+
+        # even though results are requested by raw data,
+        # Weaver still offers them with document on outputs endpoint
+        output_url = job_url + "/outputs"
+        resp = self.app.get(output_url, headers=self.json_headers)
+        assert resp.status_code == 200, "Error job outputs:\n{}".format(resp.text)
+        outputs = resp.json
+
+        self.validate_results(None, outputs, nc_data, result_links)
+
+    def test_jsonarray2netcdf_execute_async_output_by_reference_response_raw(self):
+        """
+        Jobs submitted with ``response=raw`` and single output as ``transmissionMode=reference`` must a link.
+
+        Contents should be empty, and the reference should be provided with HTTP ``Link`` header.
+
+        .. seealso::
+            https://docs.ogc.org/is/18-062r2/18-062r2.html#req_core_process-execute-sync-raw-ref
+        """
+        with contextlib.ExitStack() as stack_exec:
+            body, nc_data = self.setup_inputs(stack_exec)
+            body.update({
+                "response": ExecuteResponse.RAW,  # by value/reference important here
+                "outputs": [{"id": "output", "transmissionMode": ExecuteTransmissionMode.REFERENCE}],  # Link header
+            })
+            for mock_exec in mocked_execute_celery():
+                stack_exec.enter_context(mock_exec)
+            path = "/processes/jsonarray2netcdf/jobs"
+            resp = mocked_sub_requests(self.app, "post_json", path,
+                                       data=body, headers=self.json_headers, only_local=True)
+
+        assert resp.status_code == 201, "Error: {}".format(resp.json)
+        assert resp.content_type in ContentType.APP_JSON
+        job_url = resp.json["location"]
+        self.monitor_job(job_url, return_status=True)  # don't fetch results automatically
+
+        resp = self.app.get("{}/results".format(job_url), headers=self.json_headers)
+        assert resp.status_code < 400, "Error: {}".format(resp.json)
+        assert resp.status_code == 204, "Body should be empty since all outputs requested by reference (Link header)"
+        assert resp.content_type is None
+        assert resp.headers
+        result_links = [hdr for hdr in resp.headers if hdr[0].lower() == "link"]
+
+        # even though results are requested by Link reference,
+        # Weaver still offers them with document on outputs endpoint
+        output_url = job_url + "/outputs"
+        resp = self.app.get(output_url, headers=self.json_headers)
+        assert resp.status_code == 200, "Error job outputs:\n{}".format(resp.json)
+        outputs = resp.json
+
+        self.validate_results(None, outputs, nc_data, result_links)
+
+    def test_jsonarray2netcdf_execute_sync(self):
+        """
+        Job submitted with ``mode=sync`` or ``Prefer`` header for sync should respond directly with the results schema.
+
+        .. seealso::
+            https://docs.ogc.org/is/18-062r2/18-062r2.html#sc_execute_response
+        """
+        with contextlib.ExitStack() as stack_exec:
+            body, nc_data = self.setup_inputs(stack_exec)
+            body.update({
+                "response": ExecuteResponse.DOCUMENT,
+                "outputs": [{"id": "output", "transmissionMode": ExecuteTransmissionMode.VALUE}]
+            })
+            for mock_exec in mocked_execute_celery():
+                stack_exec.enter_context(mock_exec)
+            headers = {"Prefer": "wait=10"}
+            headers.update(self.json_headers)
+            path = "/processes/jsonarray2netcdf/jobs"
+            resp = mocked_sub_requests(self.app, "post_json", path,
+                                       data=body, headers=headers, only_local=True)
 
         assert resp.status_code == 200, "Error: {}".format(resp.json)
         assert resp.content_type in ContentType.APP_JSON
 
-        # since sync, all status details are already available!
-        assert resp.json["status"] == Status.SUCCEEDED
+        # since sync, results are directly available instead of job status
+        # even if results are returned directly (instead of status),
+        # status location link is available for reference as needed
         assert "Location" in resp.headers
-        # validate indeed sync
+        # validate sync was indeed applied (in normal situation, not considering mock test that runs in sync)
         assert resp.headers["Preference-Applied"] == headers["Prefer"]
-        # following details not available yet in async, but are in sync
-        assert isinstance(resp.json["created"], str) and resp.json["created"]
-        assert isinstance(resp.json["finished"], str) and resp.json["finished"]
-        assert isinstance(resp.json["duration"], str) and resp.json["duration"]
-        assert isinstance(resp.json["progress"], int) and resp.json["progress"] == 100
+        # following details should not be available since results are returned in sync instead of async job status
+        for field in ["status", "created", "finished", "duration", "progress"]:
+            assert field not in resp.json
 
+        # validate that job can still be found and its metadata are defined although executed in sync
         job_url = resp.headers["Location"]
+        resp = self.app.get(job_url, headers=self.json_headers)
+        assert resp.status_code == 200
+        assert resp.content_type == ContentType.APP_JSON
+        for field in ["status", "created", "finished", "duration", "progress"]:
+            assert field in resp.json
+        assert resp.json["status"] == Status.SUCCEEDED
+        assert resp.json["progress"] == 100
+
         out_url = f"{job_url}/results"
         resp = self.app.get(out_url, headers=self.json_headers)
         assert resp.status_code == 200
@@ -180,34 +362,4 @@ class BuiltinAppTest(WpsConfigBase):
         assert resp.status_code == 200, "Error job outputs:\n{}".format(resp.json)
         outputs = resp.json
 
-        self.validate_results(results, outputs, nc_data)
-
-    def validate_results(self, results, outputs, data):
-
-        # first validate format of OGC-API results
-        assert "output" in results, "Expected result ID 'output' in response body"
-        assert isinstance(results["output"], dict), "Container of result ID 'output' should be a dict"
-        assert "href" in results["output"]
-        assert "format" in results["output"]
-        fmt = results["output"]["format"]  # type: JSON
-        assert isinstance(fmt, dict), "Result format should be provided with content details"
-        assert "mediaType" in fmt
-        assert isinstance(fmt["mediaType"], str), "Result format Content-Type should be a single string definition"
-        assert fmt["mediaType"] == ContentType.APP_NETCDF, "Result 'output' format expected to be NetCDF file"
-        nc_path = results["output"]["href"]
-        assert isinstance(nc_path, str) and len(nc_path)
-        settings = get_settings_from_testapp(self.app)
-        wps_out = "{}{}".format(settings.get("weaver.url"), settings.get("weaver.wps_output_path"))
-        nc_real_path = nc_path.replace(wps_out, settings.get("weaver.wps_output_dir"))
-        assert nc_path.startswith(wps_out)
-        assert os.path.split(nc_real_path)[-1] == os.path.split(nc_path)[-1]
-        assert os.path.isfile(nc_real_path)
-        with open(nc_real_path, "r") as f:
-            assert f.read() == data
-
-        # if everything was valid for results, validate equivalent but differently formatted outputs response
-        assert outputs["outputs"][0]["id"] == "output"
-        nc_path = outputs["outputs"][0]["href"]
-        assert isinstance(nc_path, str) and len(nc_path)
-        assert nc_path.startswith(wps_out)
-        assert os.path.split(nc_real_path)[-1] == os.path.split(nc_path)[-1]
+        self.validate_results(results, outputs, nc_data, None)
