+++ weaver/wps/service.py
@@ -21,7 +21,7 @@ from weaver.owsexceptions import OWSNoApplicableCode
 from weaver.processes.convert import wps2json_job_payload
 from weaver.processes.execution import submit_job_handler
 from weaver.processes.types import ProcessType
-from weaver.processes.utils import get_job_submission_response, get_process
+from weaver.processes.utils import get_process
 from weaver.store.base import StoreProcesses
 from weaver.utils import get_header, get_registry, get_settings, get_weaver_url
 from weaver.visibility import Visibility
@@ -34,6 +34,7 @@ from weaver.wps.utils import (
     load_pywps_config
 )
 from weaver.wps_restapi import swagger_definitions as sd
+from weaver.wps_restapi.jobs.utils import get_job_submission_response
 
 LOGGER = logging.getLogger(__name__)
 if TYPE_CHECKING:
@@ -236,7 +237,7 @@ class WorkerService(ServiceWPS):
         is_workflow = proc.type == ProcessType.WORKFLOW
         tags = req.args.get("tags", "").split(",") + ["xml", "wps-{}".format(wps_request.version)]
         data = wps2json_job_payload(wps_request, wps_process)
-        body, headers = submit_job_handler(
+        resp = submit_job_handler(
             data, self.settings, proc.processEndpointWPS1,
             process_id=pid, is_local=True, is_workflow=is_workflow, visibility=Visibility.PUBLIC,
             language=wps_request.language, tags=tags, headers=dict(req.headers), context=ctx
@@ -249,11 +250,11 @@ class WorkerService(ServiceWPS):
         #   way to provide explicitly Accept header. Even our Wps1Process as Workflow step depends on this behaviour.
         accept_type = get_header("Accept", req.headers)
         if accept_type == ContentType.APP_JSON:
-            resp = get_job_submission_response(body, headers)
+            resp = get_job_submission_response(resp.body, resp.headers)
             setattr(resp, "_update_status", lambda *_, **__: None)  # patch to avoid pywps server raising
             return resp
 
-        return body
+        return resp.body
 
     @handle_known_exceptions
     def prepare_process_for_execution(self, identifier):
