+++ weaver/wps_restapi/api.py
@@ -380,11 +380,12 @@ def api_conformance(request):  # noqa: F811
         ogcapi_proc_core + "/req/core/job-results-failed",
         ogcapi_proc_core + "/req/core/job-results",
         ogcapi_proc_core + "/req/core/job-results-async-document",
+        # FIXME: support raw multipart (https://github.com/crim-ca/weaver/issues/376)
         # ogcapi_proc_core + "/req/core/job-results-async-raw-mixed-multi",
-        # ogcapi_proc_core + "/req/core/job-results-async-raw-ref",
+        ogcapi_proc_core + "/req/core/job-results-async-raw-ref",
         # ogcapi_proc_core + "/req/core/job-results-async-raw-value-multi",
-        # ogcapi_proc_core + "/req/core/job-results-async-raw-value-one",
-        # ogcapi_proc_core + "/req/core/job-results-success-sync",
+        ogcapi_proc_core + "/req/core/job-results-async-raw-value-one",
+        ogcapi_proc_core + "/req/core/job-results-success-sync",
         ogcapi_proc_core + "/req/core/job-success",
         ogcapi_proc_core + "/req/core/landingpage-op",
         ogcapi_proc_core + "/req/core/landingpage-success",
@@ -409,9 +410,10 @@ def api_conformance(request):  # noqa: F811
         ogcapi_proc_core + "/req/core/process-execute-success-async",
         ogcapi_proc_core + "/req/core/process-execute-sync-document",
         # ogcapi_proc_core + "/req/core/process-execute-sync-raw-mixed-multi",
-        # ogcapi_proc_core + "/req/core/process-execute-sync-raw-ref",
+        ogcapi_proc_core + "/req/core/process-execute-sync-raw-ref",
+        # FIXME: support raw multipart (https://github.com/crim-ca/weaver/issues/376)
         # ogcapi_proc_core + "/req/core/process-execute-sync-raw-value-multi",
-        # ogcapi_proc_core + "/req/core/process-execute-sync-raw-value-one",
+        ogcapi_proc_core + "/req/core/process-execute-sync-raw-value-one",
         ogcapi_proc_core + "/req/core/pl-limit-definition",
         ogcapi_proc_core + "/req/core/pl-limit-response",
         ogcapi_proc_core + "/req/core/process-list",
